Fix the logic of the ensure_tx_exists PR
- Move the functions to EVMTransactions since they should not be in a
DB specific class.

- Fix the optimism tuple writing logic. It should not call
write_tuples() there as it's already called in the super() call. Can
just plain old write the l1 fee with a normal executemany query

- Some misc fixes
