+++ rotkehlchen/tests/api/test_ethereum_transactions.py
@@ -213,9 +213,9 @@ def query_events(server, json, expected_num_with_grouping, expected_totals_with_
 def assert_force_redecode_txns_works(api_server: 'APIServer', hashes: Optional[list[EVMTxHash]]):
     rotki = api_server.rest_api.rotkehlchen
     get_eth_txns_patch = patch.object(
-        rotki.chains_aggregator.ethereum.transactions_decoder.dbevmtx,
+        rotki.chains_aggregator.ethereum.transactions_decoder.transactions,
         'get_or_create_transaction',
-        wraps=rotki.chains_aggregator.ethereum.transactions_decoder.dbevmtx.get_or_create_transaction,  # noqa: E501
+        wraps=rotki.chains_aggregator.ethereum.transactions_decoder.transactions.get_or_create_transaction,  # noqa: E501
     )
     get_or_decode_txn_events_patch = patch.object(
         rotki.chains_aggregator.ethereum.transactions_decoder,
