+++ rotkehlchen/tests/db/test_evmtx.py
@@ -1,5 +1,3 @@
-import pytest
-
 from rotkehlchen.chain.accounts import BlockchainAccountData
 from rotkehlchen.chain.evm.types import EvmAccount
 from rotkehlchen.data_handler import DataHandler
@@ -12,7 +10,6 @@ from rotkehlchen.tests.utils.constants import (
     ETH_ADDRESS3,
     MOCK_INPUT_DATA,
 )
-from rotkehlchen.tests.utils.ethereum import INFURA_ETH_NODE
 from rotkehlchen.tests.utils.factories import make_evm_address, make_evm_tx_hash
 from rotkehlchen.types import (
     ChainID,
@@ -129,55 +126,6 @@ def test_add_get_evm_transactions(data_dir, username, sql_vm_instructions_cb):
             has_premium=True,
         )
         assert result == [tx1, tx3]
-
-
-@pytest.mark.vcr()
-@pytest.mark.parametrize('ethereum_manager_connect_at_start', [(INFURA_ETH_NODE,)])
-@pytest.mark.parametrize('ethereum_accounts', [[ETH_ADDRESS1, ETH_ADDRESS2, ETH_ADDRESS3]])
-def test_get_or_create_transaction(ethereum_accounts, ethereum_inquirer):
-    """Tests that get_or_create_transaction works fine. By testing get_or_create_transaction
-    it checks that the requirements for a transaction are met before returning it."""
-    msg_aggregator = MessagesAggregator()
-    tx_hash = deserialize_evm_tx_hash('0x2e68dd828594ea71fe0afccfad7748c90ffcd0f1add44d0d249bf79887da9318')  # noqa: E501
-    expected_transaction = EvmTransaction(
-        tx_hash=tx_hash,
-        chain_id=ChainID.ETHEREUM,
-        timestamp=Timestamp(1451706400),
-        block_number=3,
-        from_address=ethereum_accounts[1],
-        to_address=ethereum_accounts[2],
-        value=FVal('4000000'),
-        gas=FVal('5000000'),
-        gas_price=FVal('2000000000'),
-        gas_used=FVal('25000000'),
-        input_data=MOCK_INPUT_DATA,
-        nonce=1,
-    )
-
-    with ethereum_inquirer.database.conn.read_ctx() as cursor:
-        dbevmtx = DBEvmTx(ethereum_inquirer.database)
-        # check that the transaction is properly added to the DB
-        returned_tx, tx_receipt = dbevmtx.get_or_create_transaction(cursor, ethereum_inquirer, tx_hash, relevant_address=ethereum_accounts[2])  # noqa: E501
-        assert len(msg_aggregator.consume_errors()) == 0
-        assert len(msg_aggregator.consume_warnings()) == 0
-        assert returned_tx == expected_transaction
-        assert tx_receipt is not None
-        # check that the existing transaction is properly returned from the DB
-        returned_tx, tx_receipt = dbevmtx.get_or_create_transaction(cursor, ethereum_inquirer, tx_hash, relevant_address=ethereum_accounts[2])  # noqa: E501
-        assert len(msg_aggregator.consume_errors()) == 0
-        assert len(msg_aggregator.consume_warnings()) == 0
-        assert returned_tx == expected_transaction
-
-    # check that if there is a tx with no receipt, the receipt is created because it is a
-    # requirement for an evm tx.
-    with ethereum_inquirer.database.conn.write_ctx() as write_cursor:
-        write_cursor.execute('DELETE FROM evmtx_receipts WHERE tx_id = ?', (returned_tx.db_id,))
-
-    with ethereum_inquirer.database.conn.read_ctx() as cursor:
-        tx_receipt = dbevmtx.get_receipt(cursor, tx_hash, ethereum_inquirer.chain_id)
-        assert tx_receipt is None
-        returned_tx, tx_receipt = dbevmtx.get_or_create_transaction(cursor, ethereum_inquirer, tx_hash, relevant_address=ethereum_accounts[2])  # noqa: E501
-        assert tx_receipt is not None
 
 
 def test_query_also_internal_evm_transactions(data_dir, username, sql_vm_instructions_cb):
